Rename "guess" state to "tentativeGuess"
The "guess" state variable in GuessInput and Game components has been renamed to "tentativeGuess" for better clarity. All references to this state variable in the code have also been updated accordingly. Now it's easier to understand that this variable holds the current attempt that hasn't been committed as a guess yet.